tmux: use a different way to save buffer
More recent versions of tmux allow to name the buffer we're saving to
instead of just an index. Older versions (1.8) which just use a buffer
index, don't create a new buffer with that index. The buffer already
needs to exist.

To make this work with newer and older versions of tmux, we use '-p' to
print the contents to stdout out, capture it and write it to a file
ourselfs.