tests: compare against Bytestring instead of file
Instead of writing a file to the default tempdir, I think it's easier to
handle a comparison with a Bytestring. That means, we still write a file
to our test directory (which we'll clean up after a successful session),
but that means we can easier compartmentalize the whole thing rather
then keeping track of stray output files.

Issue: #38