Poll for changed terminal output
We need to wait until the application has processed events and rendered
the changes. Having the expectation of an instance change in the
terminal after all keys are send is probably not realistic and will lead
to unreliable tests.

Furthermore, it saves us to maintain the whole over-engineered ping-pong
socket sending code, which doesn't really help here.

Issue: #38